Add file module to export js files
